add test for precompiled vm tests
